refactor: modularize application structure
